Conflicts:
	build.gradle
	src/main/java/li/cil/tis3d/client/render/TextureLoader.java
	src/main/java/li/cil/tis3d/common/ProxyCommon.java
	src/main/java/li/cil/tis3d/common/TIS3D.java
	src/main/java/li/cil/tis3d/common/entity/EntityInfraredPacket.java
	src/main/java/li/cil/tis3d/common/module/ModuleDisplay.java
	src/main/java/li/cil/tis3d/common/module/ModuleInfrared.java
	src/main/java/li/cil/tis3d/common/module/ModuleRandom.java
	src/main/java/li/cil/tis3d/common/module/ModuleRedstone.java
	src/main/java/li/cil/tis3d/common/tile/TileEntityCasing.java

+++ src/main/java/li/cil/tis3d/api/InfraredAPI.java
@@ -20,7 +20,7 @@ public final class InfraredAPI {
      * @param value     the value the packet carries.
      * @return the packet that was spawned.
      */
-    public static InfraredPacket sendPacket(final World world, final Vec3 position, final Vec3 direction, final int value) {
+    public static InfraredPacket sendPacket(final World world, final Vec3 position, final Vec3 direction, final short value) {
         if (API.infraredAPI != null)
             return API.infraredAPI.sendPacket(world, position, direction, value);
         return null;
