+++ src/main/java/li/cil/tis3d/common/module/execution/instruction/InstructionBitwiseXorImmediate.java
@@ -4,9 +4,9 @@ import li.cil.tis3d.common.module.execution.Machine;
 import li.cil.tis3d.common.module.execution.MachineState;
 
 public final class InstructionBitwiseXorImmediate implements Instruction {
-    private final int value;
+    private final short value;
 
-    public InstructionBitwiseXorImmediate(final int value) {
+    public InstructionBitwiseXorImmediate(final short value) {
         this.value = value;
     }
