+++ src/main/java/li/cil/tis3d/common/module/execution/instruction/InstructionSwap.java
@@ -9,7 +9,7 @@ public final class InstructionSwap implements Instruction {
     @Override
     public void step(final Machine machine) {
         final MachineState state = machine.getState();
-        final int tmp = state.acc;
+        final short tmp = state.acc;
         state.acc = state.bak;
         state.bak = tmp;
         state.pc++;
